+++ Helper/Data.php
@@ -47,7 +47,7 @@ class Data extends AbstractHelper
     const MODULE_NAME = 'Nosto_Cmp';
 
     /** @var NostoHelperScope */
-    private $nostoHelperScope;
+    private NostoHelperScope $nostoHelperScope;
 
     /**
      * Path to the configuration object that stores category sorting
@@ -70,7 +70,7 @@ class Data extends AbstractHelper
     const XML_PATH_CATEGORY_MAX_PRODUCT_LIMIT = 'nosto_cmp/limit/max_products';
 
     /** @var ModuleListInterface */
-    private $moduleList;
+    private ModuleListInterface $moduleList;
 
     /**
      * Data constructor.
@@ -138,7 +138,7 @@ class Data extends AbstractHelper
      * @param Store|null $store
      * @return mixed|null
      */
-    public function getStoreConfig($path, Store $store = null)
+    public function getStoreConfig(string $path, Store $store = null)
     {
         if ($store === null) {
             $store = $this->nostoHelperScope->getStore(true);