+++ Model/Service/Facet/BuildWebFacetService.php
@@ -62,19 +62,19 @@ class BuildWebFacetService
     }
 
     /** @var State */
-    private $state;
+    private State $state;
 
     /** @var NostoCategoryBuilder */
-    private $nostoCategoryBuilder;
+    private NostoCategoryBuilder $nostoCategoryBuilder;
 
     /** @var CategoryRepository */
-    private $categoryRepository;
+    private CategoryRepository $categoryRepository;
 
     /** @var NostoHelperData */
-    private $nostoHelperData;
+    private NostoHelperData $nostoHelperData;
 
     /** @var string */
-    private $brand;
+    private string $brand;
 
     /**
      * BuildWebFacetService constructor.
@@ -150,7 +150,7 @@ class BuildWebFacetService
         $filter = $item->getFilter();
 
         if ($filter instanceof Category) {
-            $categoryId = $item->getData('value');
+            $categoryId = (int)$item->getData('value');
             $category = $this->getCategoryName($store, $categoryId);
             if ($category == null) {
                 $this->trace('Could not get category from filters');
@@ -217,7 +217,7 @@ class BuildWebFacetService
      * @return string|null
      * @throws NoSuchEntityException
      */
-    private function getCategoryName(Store $store, $categoryId): ?string
+    private function getCategoryName(Store $store, int $categoryId): ?string
     {
         /**
          * Argument is of type \Magento\Catalog\Api\Data\CategoryInterface
@@ -238,8 +238,8 @@ class BuildWebFacetService
      */
     private function mapValueToFilter(IncludeFilters &$includeFilters, Store $store, string $name, $value)
     {
-        if ($this->brand == null) {
-            $this->brand = $this->nostoHelperData->getBrandAttribute($store);
+        if (empty($this->brand)) {
+            $this->brand = $this->nostoHelperData->getBrandAttribute($store) ?? '';
         }
 
         switch (strtolower($name)) {